requirements.txt
scipy
tensorflow
tflearn
nltk
django==2.1.3
discord.py
